requirements.txt
# Core dependencies for data processing pipeline
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
structlog>=22.1.0

# Kafka
confluent-kafka>=1.9.0

# Database clients
asyncpg>=0.28.0
redis>=4.3.0
clickhouse-driver>=0.2.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Observability
prometheus-client>=0.17.0
psutil>=5.9.0
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-exporter-otlp>=1.15.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.991
flake8>=5.0.0

# Utilities
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

